Add files via upload
First off, your output is far superior - it includes formatting and one row per game. A couple of notes/questions for discussions:
1) Add column for time of game to potentially use as a feature in future analytics
2) ID should be in format of YYYY-MM-DD_ATEAM_HTEAM so that when sorted, it will sort by date accurately rather than clubbing months together
3) Add time of scrape to the ID? I'm thinking about after we join the odds to these games, should we be overwriting or adding rows with the same game ID but updated time of scrape and updated odds or winning percentage
4) When checking for games to bet on, should we only bet on current day games? One could argue that we only want to bet on current day games because various factors of the game may change leading up to the actual game (e.g., starting pitcher gets sick/subbed). One could also argue that we want to take advantage of better odds leading up to the actual game, so betting a couple of days out would also be beneficial.

I will try to work on committing changes (is that how you say it?) to incorporate 1 and 2. We can discuss all the items above whenever.

Just uploading my code in case you see something that could be used from it.